Argon version 1.1 (tweaks)
Number of slices in ShuffleSlices is a tunable parameter to fix the
degree of parallelism available to the adversary.
Number of slices  and the tag length are now part of the input to
further prevent the adversary from amortizing exhaustive search.
Associated data $X$ can now be processed as part of the input. It can be
an arbitrary string, containing application context, user data, or any
other parameters that shall be binded to the hash value.
All the inputs are hashed with a cryptographic hash function (Blake2b)
in order to extract the entropy from the password and the salt as soon
as possible and put it into a short string $I$.
Due to the use of a hash function, password and salt may have arbitrary
(up to $2^{32}$ bytes) length.
The string $I$ is partitioned in only 4 parts, which fill the first 8
bytes of each 16-byte memory block, whereas the rest is taken by the
counter. This has been done for simplicity.
The Shuffle algorithm now employs a new formula to compute the value of
the counter $j$, which should not allow to parallelize the algorithm, as
it is the case for RC4.
To produce the tag, memory is divided in two, then both parts are XORed
into two 128-bit blocks. Both blocks are then input to a variable length
hash function (Blake2b),   which produces tags of arbitrary length.